requirement.txt
playwright==1.52.0
python-dotenv==1.1.0
behave==1.2.6
allure-behave==2.13.5
pre-commit
opencv-python==4.12.0.88